requirements/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.10.1/constraints-3.11.txt"

apache-airflow-providers-snowflake==5.7.0
apache-airflow-providers-mysql==5.7.0
google-cloud-bigquery
google-cloud-pubsub
google-auth

# Other Libraries
requests
Faker